feat: implement multi-ring support and format synchronization in ESP32
   - Update NAV reader to support uint16 ring counts (>255 rings)
   - Synchronize ESP32 structures with updated NAV binary format

+++ lib/maps/src/nav_reader.hpp
@@ -118,7 +118,7 @@ struct NavFeature
     NavCoord* coords = nullptr;
     uint16_t coordCount = 0;
     uint16_t* ringEnds = nullptr;
-    uint8_t ringCount = 0;
+    uint16_t ringCount = 0;
     
     // Pixel offset of the tile top-left relative to viewport top-left
     int16_t tilePixelOffsetX;
